requirements.txt
fastapi[all] 
motor 
bcrypt 
pyjwt 
python-dotenv